# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quick Turn",
        "description": "Documentación de API QuickTurn",
        "contact": {
            "email": "germanca8@correo.ugr.es"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/seccions": {
            "get": {
                "tags": [
                    "Secciones"
                ],
                "summary": "Lista todas las secciones",
                "description": "Endpoint para listar las secciones",
                "operationId": "00d9d7129146c7a604a5be6ec4a6868a",
                "responses": {
                    "200": {
                        "description": "Lista todas las secciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Seccion"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Secciones"
                ],
                "summary": "Crea una seccion",
                "description": "Endpoint para almacenar una seccion",
                "operationId": "ee1a1facee020ed58a95693b24abab6d",
                "requestBody": {
                    "$ref": "#/components/requestBodies/StoreSeccion"
                },
                "responses": {
                    "200": {
                        "description": "Seccion creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Seccion"
                                }
                            }
                        }
                    },
                    "444": {
                        "$ref": "#/components/responses/444"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/seccions/{seccionID}": {
            "get": {
                "tags": [
                    "Secciones"
                ],
                "summary": "Muestra una seccion por ID",
                "description": "Endpoint para mostrar una seccion",
                "operationId": "850451b466dff0c6ea019d1f23b5075f",
                "parameters": [
                    {
                        "name": "seccionID",
                        "in": "path",
                        "description": "ID de la seccion a obtener",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la seccion",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Seccion"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Secciones"
                ],
                "summary": "Elimina una seccion",
                "description": "Endpoint para eliminar una seccion",
                "operationId": "aa4efa7b498eeba7538041e8efff96b6",
                "parameters": [
                    {
                        "name": "seccionID",
                        "in": "path",
                        "description": "ID de la seccion a eliminar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Seccion eliminada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "deleted": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Secciones"
                ],
                "summary": "Actualiza una seccion",
                "description": "Endpoint para actualizar una seccion",
                "operationId": "7fb08100a6563394fc35c452f827fb49",
                "parameters": [
                    {
                        "name": "seccionID",
                        "in": "path",
                        "description": "ID de la seccion a actualizar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateSeccion"
                },
                "responses": {
                    "200": {
                        "description": "Seccion actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Seccion"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/seccions/{seccion}/aumentarTurnoActual": {
            "post": {
                "tags": [
                    "Trabajadores"
                ],
                "summary": "Aumenta el turno actual de la sección",
                "description": "Endpoint para aumentar el turno actual de la sección",
                "operationId": "54f1fee6fa060284a70564cf1f81c8dc",
                "parameters": [
                    {
                        "name": "seccion",
                        "in": "path",
                        "description": "ID de la seccion a actualizar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Turno aumentado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Seccion"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                }
            }
        },
        "/api/seccions/{seccion}/verTurnoActual": {
            "get": {
                "tags": [
                    "Trabajadores"
                ],
                "summary": "Imprime el turno actual de la sección",
                "description": "Endpoint para ver el turno actual de la sección",
                "operationId": "bd4064589ae2338d259b7091ee3ef304",
                "parameters": [
                    {
                        "name": "seccion",
                        "in": "path",
                        "description": "ID de la seccion de la que quiero el turno",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Turno impreso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Seccion"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                }
            }
        },
        "/api/seccions/{entidadID}": {
            "post": {
                "tags": [
                    "Secciones"
                ],
                "summary": "Lista todas las secciones de 1 entidad",
                "description": "Endpoint para mostrar todas las secciones asociadas a una entidad",
                "operationId": "f26d05de4f28df05c40eef7a167e963d",
                "parameters": [
                    {
                        "name": "entidadID",
                        "in": "path",
                        "description": "ID de la entidad",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista todas las secciones de 1 entidad",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Seccion"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Seccion": {
                "properties": {
                    "id": {
                        "description": "Id asociado",
                        "type": "string",
                        "example": "670b9562-b30d-52d5-b827-655787665500"
                    },
                    "entidad_id": {
                        "description": "Id de la entidad",
                        "type": "string",
                        "example": "670b9562-b30d-52d5-b827-655787665500"
                    },
                    "nombreSeccion": {
                        "description": "Nombre de la seccion",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Mercadona"
                    },
                    "turnoActual": {
                        "description": "Numero del turno actual",
                        "type": "integer",
                        "maxLength": 52,
                        "example": "363"
                    },
                    "ultimoTurno": {
                        "description": "Numero del ultimo turno",
                        "type": "integer",
                        "maxLength": 52,
                        "example": "363"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "example": "2022-03-15T13:23:02.000000Z"
                    },
                    "updated_at": {
                        "description": "Fecha de última modificación",
                        "type": "string",
                        "example": "2022-03-15T13:23:02.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "404": {
                "description": "No query results for model",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "444": {
                "description": "Name its already used",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Name is already used"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "requestBodies": {
...
            "StoreSeccion": {
                "description": "Seccion que necesita ser creada",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "entidad_id",
                                "nombreSeccion"
                            ],
                            "properties": {
                                "entidad_id": {
                                    "$ref": "#/components/schemas/Seccion/properties/entidad_id"
                                },
                                "nombreSeccion": {
                                    "$ref": "#/components/schemas/Seccion/properties/nombreSeccion"
                                },
                                "turnoActual": {
                                    "$ref": "#/components/schemas/Seccion/properties/turnoActual"
                                },
                                "ultimoTurno": {
                                    "$ref": "#/components/schemas/Seccion/properties/ultimoTurno"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "UpdateSeccion": {
                "description": "Seccion que necesita ser actualizada",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "nombreSeccion"
                            ],
                            "properties": {
                                "nombreSeccion": {
                                    "$ref": "#/components/schemas/Seccion/properties/nombreSeccion"
                                },
                                "turnoActual": {
                                    "$ref": "#/components/schemas/Seccion/properties/turnoActual"
                                },
                                "ultimoTurno": {
                                    "$ref": "#/components/schemas/Seccion/properties/ultimoTurno"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Laravel Sanctum Token Auth",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}